improving wasm API that returns evaluation metrics
