--- packages/lib-wasm/src/processing/evaluator/aggregate_counts_stats.rs
@@ -1,85 +1,0 @@
-use super::stats::{
-    calc_fabricated_combinations_percentage, calc_mean_combinations_count_error,
-    calc_mean_combinations_count_error_by_len, calc_mean_combinations_length_by_bucket,
-    calc_mean_proportional_error, calc_mean_proportional_error_by_bucket,
-    calc_original_mean_combinations_count, calc_original_mean_combinations_count_by_len,
-    calc_suppressed_combinations_percentage,
-};
-use crate::{
-    processing::aggregator::aggregate_result::WasmAggregateResult,
-    utils::js::ts_definitions::{JsAggregateCountsStatistics, JsResult},
-};
-use sds_core::processing::{aggregator::typedefs::AggregatedMetricByLenMap, evaluator::Evaluator};
-use serde::Serialize;
-use wasm_bindgen::{prelude::wasm_bindgen, JsCast, JsValue};
-
-#[wasm_bindgen]
-#[derive(Serialize, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct WasmAggregateCountsStatistics {
-    suppressed_combinations_percentage: f64,
-    fabricated_combinations_percentage: f64,
-    original_mean_combinations_count: f64,
-    original_mean_combinations_count_by_len: AggregatedMetricByLenMap,
-    mean_combinations_count_error: f64,
-    mean_combinations_count_error_by_len: AggregatedMetricByLenMap,
-    mean_proportional_error: f64,
-    mean_proportional_error_by_bucket: AggregatedMetricByLenMap,
-    mean_combinations_length_by_bucket: AggregatedMetricByLenMap,
-}
-
-#[wasm_bindgen]
-impl WasmAggregateCountsStatistics {
-    #[wasm_bindgen(constructor)]
-    pub fn from_aggregate_results(
-        sensitive_aggregate_result: &WasmAggregateResult,
-        reportable_aggregate_result: &WasmAggregateResult,
-        resolution: usize,
-    ) -> JsResult<WasmAggregateCountsStatistics> {
-        let preservation_by_count = Evaluator::default().calc_preservation_by_count(
-            sensitive_aggregate_result,
-            reportable_aggregate_result,
-            resolution,
-        );
-
-        Ok(WasmAggregateCountsStatistics {
-            suppressed_combinations_percentage: calc_suppressed_combinations_percentage(
-                sensitive_aggregate_result,
-                reportable_aggregate_result,
-            ),
-            fabricated_combinations_percentage: calc_fabricated_combinations_percentage(
-                sensitive_aggregate_result,
-                reportable_aggregate_result,
-            ),
-            original_mean_combinations_count: calc_original_mean_combinations_count(
-                sensitive_aggregate_result,
-            ),
-            original_mean_combinations_count_by_len: calc_original_mean_combinations_count_by_len(
-                sensitive_aggregate_result,
-            ),
-
-            mean_combinations_count_error: calc_mean_combinations_count_error(
-                sensitive_aggregate_result,
-                reportable_aggregate_result,
-            ),
-            mean_combinations_count_error_by_len: calc_mean_combinations_count_error_by_len(
-                sensitive_aggregate_result,
-                reportable_aggregate_result,
-            ),
-            mean_proportional_error: calc_mean_proportional_error(&preservation_by_count),
-            mean_proportional_error_by_bucket: calc_mean_proportional_error_by_bucket(
-                &preservation_by_count,
-            ),
-            mean_combinations_length_by_bucket: calc_mean_combinations_length_by_bucket(
-                &preservation_by_count,
-            ),
-        })
-    }
-
-    #[wasm_bindgen(js_name = "toJs")]
-    pub fn to_js(&self) -> JsResult<JsAggregateCountsStatistics> {
-        Ok(JsValue::from_serde(self)
-            .map_err(|err| JsValue::from(err.to_string()))?
-            .unchecked_into::<JsAggregateCountsStatistics>())
-    }
-}
