+++ packages/lib-wasm/src/processing/evaluator/mod.rs
@@ -1,9 +1,5 @@
 mod stats;
 
-pub mod aggregate_counts_stats;
-
 pub mod evaluate_result;
 
-pub mod sensitive_data_stats;
-
-pub mod synthetic_data_stats;
+pub mod microdata_data_stats;
