+++ packages/lib-wasm/src/processing/sds_context.rs
@@ -175,6 +175,7 @@ impl SDSContext {
             &self.reportable_aggregate_result,
             &self.synthetic_aggregate_result,
             self.resolution,
+            reporting_length,
         )?;
 
         self.clear_navigate();
